Fixes
* getdents is more compliant to POSIX
* basic user data validation
* fix bug causing returning pages in wrong order from page_range_get

Improvements
* add syscall lib function; all syscalls in userspace are using it
* errno and perror
* fix overwritten eax during exception with error code
* improved mount; introduced file_system.mount operation instead of
  read_super; removed read_inode
* better readdir
* write in file_operations takes const char* instead of char*
* don't allocate additional page for args
* some clean-up in string.[ch]
* better logging (log debug takes flag as first argument)
* create kernel/page.h
* slab allocator which replaced kmalloc; removed `new` macro; some
  improvements in fork.c
* save user esp to the context.esp2 during syscall/irq/exc
* better signal handling; do not wake process when it receives a signal
* no additional stack is allocated for signal handling

+++ kernel/lib/string.c
@@ -63,7 +63,7 @@ char* strrchr(const char* string, int c)
 #endif
 
 #ifndef __HAVE_ARCH_STRLEN
-unsigned int strlen(const char* string)
+size_t strlen(const char* string)
 {
     char* temp;
     for (temp = (char*)string; *temp != 0; temp++);
@@ -72,7 +72,7 @@ unsigned int strlen(const char* string)
 #endif
 
 #ifndef __HAVE_ARCH_STRNLEN
-unsigned int strnlen(const char* s, size_t maxlen)
+size_t strnlen(const char* s, size_t maxlen)
 {
     register const char* e;
     size_t n;
